Refactor tests for retention readers
